requirements.txt
# ZeroGPU - MUST be first for decorator detection
spaces>=0.28.3

# FastAPI for backend
fastapi>=0.100.0
uvicorn>=0.23.0

# Gradio 5 for HuggingFace Spaces
gradio>=5.0.0
huggingface-hub>=0.25.0
pydantic==2.10.6

# Core
python-dotenv>=1.0.0

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# RAG
chromadb>=0.4.0
openai>=1.0.0
tiktoken>=0.5.0
pinecone>=3.0.0

# Model (for fine-tuned Qwen)
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
accelerate>=0.20.0
bitsandbytes>=0.46.1

# Utilities
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
python-frontmatter>=1.0.0
sentence-transformers>=2.2.0
